pyproject.toml
[project]
name = "ccmd-dashboard"
version = "0.1.0"
description = "CCMD Media Intelligence Dashboard — OSW/ICC unclassified prototype"
readme = "README.md"
authors = [
    { name = "OSW Prototype" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18",
    "pydantic-settings>=2.13",
    "pyyaml>=6.0",
    "sqlmodel>=0.0.38",
    "typer>=0.24",
]

[project.optional-dependencies]
ingest = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "trafilatura>=1.12",
    "python-dateutil>=2.9",
]
classify = [
    "spacy>=3.7",
    "sentence-transformers>=3.0",
    "anthropic>=0.40",
    "apscheduler>=3.10",
]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "weasyprint>=62",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.7",
]

[project.scripts]
dashboard = "ccmd_dashboard.cli:app"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
